+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+### Changed
+- much improved timezone keyword search
+
 ### Fixed
 - numeric timezone abbreviations (e. g. "+0845") were ripped apart
 - slider buttons in Settings not always showing correct state immediately after reset
@@ -47,7 +50,7 @@
 - "What's New" under Settings (references CHANGELOG.md on GitHub main branch, = this file)
 
 ### Changed
-- up tp 4 custom tabs rather than 3
+- up to 4 custom tabs rather than 3
 - GMST (and LMST) expressed in time units/hours as well as angular units/degrees
 
 ### Fixed